fix issues and add functionality for deactivating the user

+++ DailyTaskMaker.Infrastructure/DataModels/UserDetailsModel.cs
@@ -15,7 +15,7 @@ namespace DailyTaskMakerAPI.Models
         public  string? EmailId { get; set; }
 
         
-        public int ? MobileNumber { get; set; }
+        public long ? MobileNumber { get; set; }
 
         
         public  string ?  DisplayName { get; set; }
